chore(workmanagement): consolidate listeners, memoize grouping, extract kanban CSS, add group util & tests, harden undo/DnD updates
