+++ src/pages/workmanagement.css
@@ -1,0 +1,72 @@
+/* Kanban/DnD styles extracted from WorkManagementPage.js */
+/* カード装飾 */
+.kanban-card {
+    background: #fff;
+    border-radius: 0.75rem;
+    box-shadow: 0 2px 8px rgba(0,0,0,0.06);
+    border: 1.5px solid #e5e7eb;
+    transition: box-shadow 0.2s, border 0.2s;
+    margin-bottom: 0;
+}
+.kanban-card:hover {
+    box-shadow: 0 4px 16px rgba(99,102,241,0.10);
+    border-color: #6366f1;
+}
+.kanban-card.dragging {
+    opacity: 0.5;
+    cursor: grabbing;
+    transform: scale(1.05);
+    box-shadow: 0 8px 24px rgba(99,102,241,0.18);
+    border-color: #6366f1;
+    background: #f1f5f9;
+}
+.drag-over {
+    background-color: #e0e7ff !important;
+    border-radius: 0.75rem;
+    box-shadow: 0 0 0 2px #6366f1;
+}
+.drop-placeholder {
+    background-color: #c7d2fe;
+    border-radius: 0.75rem;
+    box-shadow: 0 8px 20px rgba(0,0,0,0.2);
+    border: 2px dashed #6366f1;
+    margin: 2px 0;
+    min-height: 48px;
+    height: auto;
+    padding: 1rem;
+    width: auto;
+    max-width: 100%;
+    display: flex;
+    flex-direction: column;
+    flex: none;
+    align-self: stretch;
+    box-sizing: border-box;
+    opacity: 0.5;
+    transition: all 0.15s cubic-bezier(0.4,0,0.2,1);
+    pointer-events: none;
+}
+/* カラム装飾 */
+.kanban-column {
+    background: #f8fafc;
+    border-radius: 1rem;
+    border: 1.5px solid #e5e7eb;
+    box-shadow: 0 2px 8px rgba(0,0,0,0.04);
+    padding: 1.2rem 1rem 1rem 1rem;
+    margin-bottom: 1.5rem;
+    transition: box-shadow 0.2s, border 0.2s;
+}
+.kanban-column:hover {
+    box-shadow: 0 4px 16px rgba(99,102,241,0.08);
+    border-color: #6366f1;
+}
+/* スクロールバー装飾 */
+.card-list-container::-webkit-scrollbar {
+    width: 8px;
+}
+.card-list-container::-webkit-scrollbar-thumb {
+    background: #e0e7ff;
+    border-radius: 4px;
+}
+.card-list-container::-webkit-scrollbar-track {
+    background: #f1f5f9;
+}
